move pending org logic to its own dir
