+++ frontend/src/app/admin/PendingOrganizations/PendingOrganizationsList.scss
